refactor: Ladder이름을 RandomLadder로 변경
기존의 Ladder는 사다리를 랜덤으로 만들어 준다. 이후에 등장한 사용자 지정 Ladder와의 대비시 명확한 이해를 위해서 이름변경

또한 interface 추출을 위해 RandomLadder와 UserDefineLadder모두 ladder 페키지로 이동시켰다.
